Centralize RangeOptions and ensure types are declared before use
Typescript was doing some odd behavior where in splits it was
stealing dividend's RangeOptions. Moreso, it was setting up a wierd
absoulte path to the file ((../../../Users/User/Repros/IEX-service/...)
Couldn't seem to solve it, perhaps a bug with a type deduplication?
Instead pulled the RangeOptions out into a file.